% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs.R
\docType{package}
\name{wfs}
\alias{wfs}
\title{wfs: Extracting and Reshaping World Fertility Survey Data in R}
\description{
This package facilitates the use of World Fertility Survey (WFS) datasets
made available through the Demographic and Health Surveys (DHS) data archive.
}
\section{Details}{

It includes functions to
\itemize{
\item{view the list of datasets available and the documentation,}
\item{extract data directly from the data archive or local files,}
\item{reshape the union and birth histories from wide to long
format and back, and}
\item{make local copies of the data and dictionary files.}
}
To learn more about wfs, start with the vignette (to come)
}

