% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_copy.R
\name{wfs_copy}
\alias{wfs_copy}
\title{Make local copies}
\usage{
wfs_copy(dataset, target = "")
}
\arguments{
\item{dataset}{A required string with the dataset name, e.g. "cosr02"}

\item{target}{An optional string with name of folder where to save the files,
defaults to current working directory}
}
\description{
Downloads the dictionary and data files for a dataset and saves local copies
in a specified folder or the working directory
}
\examples{
wfs_copy("cosr02")
}
