% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_extract.R
\name{wfs_extract}
\alias{wfs_extract}
\title{Data extraction}
\usage{
wfs_extract(varlist, dataset, source = "", convert.factors = TRUE)
}
\arguments{
\item{varlist}{A required string with a comma-separated
list of variables, for example "v010, v110". May include
ranges, wilcards or keywords as explained under Details}

\item{dataset}{A required string with the name of a dataset,
for example "cosr02"}

\item{source}{An optional string specifying a local folder where
to find the files, leave blank to read the files from the DHS
data archive}

\item{convert.factors}{An optional boolean, use value labels to
create factors?}
}
\value{
a data frame with attributes.
}
\description{
Extract a set of variables from a dataset reading it from
the DHS data archive or the local file system
}
\section{Details}{

The \code{varlist} is a comma-separated list of variable names, all in
lowercase (even if dictionaries use uppercase).
The list may include a \emph{range} such as  \code{v701-v705}
to extract variables v701 to v705.
It may also include the \emph{wildcards} \code{?} and \code{*}
to match one or more characters, so \code{m??2} extracts
the date of union for all unions.
You may also use the \emph{keywords} \code{unions} to extract
the union history, \code{births} for the birth history, or
\code{all} for all variables.

The \code{dataset} is required and must be the name of a
dataset, for example \code{cosr02}. The dataset consists of
a dictionary file with extension \code{.dct} and an ASCII
data file with extension \code{.dat}.

The \code{source} is the name of a local folder where the
two files mentioned above may be found. If left blank the
function will download the dictionary and data files directly
from the DHS data archive.

By default \code{convert.factors} is \code{TRUE} and we
convert a variable to a factor if the dictionary specifies
value labels and \emph{all} values in the data other than
\code{NA} have a corresponding value label. The conversion
may be turned off by setting the flag to \code{FALSE}

The function returns a data frame.

Each variable with value labels has a \code{"labels"} attribute
with the value labels, unless it was converted to a factor.
This information can be used to convert a variable to a factor
at a later time using the function \code{labelled::to_factor()}

Dictionaries may specify missing values and special codes.
We recode all missing values to \code{NA}.
For variables that are not converted to factors, we add the
special code, if any, as a \code{"special"} attribute of the variable.
Any numeric values greater or equal to the special code require
special treatment in analysis.
}

\section{Example}{

wfs_extract("v011, v111", "cosr02")
}

