% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_reshape.R
\name{wfs_reshape}
\alias{wfs_reshape}
\title{Reshaping history data}
\usage{
wfs_reshape(data, direction, history, drop = TRUE)
}
\arguments{
\item{data}{Required data frame based on a WFS standard recode file}

\item{direction}{Required string set to "long" or "wide"}

\item{history}{Required string set to "unions" or "births"}

\item{drop}{Optional boolean indicating whether to drop empty entries}
}
\value{
A reshaped data frame with attributes
}
\description{
Reshape the union or birth history from wide to long format and back.
}
\section{Details}{

\code{wfs_reshape} relies on R's \code{reshape} to do the work,
generating the appropriate function call to go from wide to long format.
Note that all history variables must have been extracted, usually by
specifying the \code{unions} or \code{births} keyword.

Typically you will first reshape in the "long" direction to create
one record per union or per birth. By default the function keeps only
actual entries in the history, but setting \code{drop = FALSE} will keep
them all.

If you reshape back to "wide" format we rely on the fact that R's
\code{reshape} is reversible, checking only that the dataset is in the
appropriate long format. By default the restored file will have only
ever-married women or mothers unless empty entries were retained.
}

