% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_view.R
\name{wfs_view}
\alias{wfs_view}
\title{View documentation}
\usage{
wfs_view(dataset = "", source = "")
}
\arguments{
\item{dataset}{String with optional dataset name, e.g. "cosr02"}

\item{source}{String with source folder or blank for DHS archive}
}
\description{
When called with no arguments this function points your browser
to the WFS home page in the DHS data archive. When called with
a dataset name it opens the data dictionary, reading it from
a local file or from the data archive.
}
\examples{
wfs_view()
wfs_view("cosr02")
}
